[Add]Tokenのデコード処理を書いた
