+++ src/main/kotlin/com/approvers/devlazaApi/controller/AuthController.kt
@@ -1,52 +1,76 @@
 package com.approvers.devlazaApi.controller
 
+import com.approvers.devlazaApi.errors.BadRequest
 import com.approvers.devlazaApi.errors.NotFound
 import com.approvers.devlazaApi.model.AuthPoster
 import com.approvers.devlazaApi.model.Token
 import com.approvers.devlazaApi.model.User
 import com.approvers.devlazaApi.repository.TokenRepository
 import com.approvers.devlazaApi.repository.UserRepository
+import com.auth0.jwt.JWT
+import com.auth0.jwt.JWTVerifier
+import com.auth0.jwt.algorithms.Algorithm
+import com.auth0.jwt.exceptions.JWTVerificationException
+import com.auth0.jwt.interfaces.DecodedJWT
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
+import java.io.UnsupportedEncodingException
 import java.util.*
 import javax.validation.Valid
 
 @RestController
 @RequestMapping("/auth")
 class AuthController(
-		private val tokenRepository: TokenRepository,
-		private val userRepository: UserRepository
+        private val tokenRepository: TokenRepository,
+        private val userRepository: UserRepository
 ) {
-	private val notFound = User(
-				name="User not found.",
-				passWord="",
-				showId="",
-				mailAddress=""
-		)
+    @PostMapping("/")
+    fun getUserInfo(@Valid @RequestBody authPoster: AuthPoster): User {
+        val token: String = authPoster.token
 
-	@PostMapping("/")
-	fun getUserInfo(@Valid @RequestBody authPoster: AuthPoster): User {
-		val token: String = authPoster.token
+        val userId: UUID = tokenRepository.getUserIdFromToken(token)
 
-		val userId: UUID = tokenRepository.getUserIdFromToken(token)
+        val tmp: List<User> = userRepository.findById(userId)
 
-		val tmp: List<User> = userRepository.findById(userId)
+        if (tmp.isEmpty()) throw NotFound("No users were found for that token")
 
-		if (tmp.isEmpty()) return notFound
+        return tmp[0]
+    }
 
-		return tmp[0]
-	}
+    private fun TokenRepository.getUserIdFromToken(token: String): UUID{
+        val checkedToken: Token = this.checkToken(token)
+        return checkedToken.userId
+    }
 
-	private fun TokenRepository.getUserIdFromToken(token: String): UUID{
-		val checkedToken: Token = this.checkToken(token)
-		return checkedToken.userId
-	}
+    private fun TokenRepository.checkToken(token: String): Token{
+        val tokenList: List<Token> = this.findByToken(token)
+        if (tokenList.isEmpty()) throw NotFound("This token is invalid")
+        return tokenList[0]
+    }
 
-	private fun TokenRepository.checkToken(token: String): Token{
-		val tokenList: List<Token> = this.findByToken(token)
-		if (tokenList.isEmpty()) throw NotFound("This token is invalid")
-		return tokenList[0]
-	}
+    private val secret: String = System.getenv("secret") ?: "secret"
+
+    private fun decode(token: String): UUID{
+        val userId: UUID
+        try{
+            val algorithm: Algorithm = Algorithm.HMAC256(secret)
+            val verifier: JWTVerifier = JWT.require(algorithm).build()
+
+            val decodedJWT: DecodedJWT = verifier.verify(token)
+
+            userId = UUID.fromString(
+                    decodedJWT.getClaim("USER_ID").asString()
+            )
+        }catch (e: Exception){
+            when(e){
+                is UnsupportedEncodingException, is JWTVerificationException
+                -> throw BadRequest("token is invalid")
+                else -> throw e
+            }
+        }
+
+        return userId
+    }
 }
