+++ src/main/kotlin/com/approvers/devlazaApi/model/User.kt
@@ -11,57 +11,57 @@ import javax.validation.constraints.Email
 
 @Entity
 data class User(
-		@Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
-		@Column(name="name", nullable=false) var name: String,
-		@Column(name="birthDay", nullable=false) var birthDay: LocalDateTime? = LocalDateTime.of(1900, 1, 1, 0, 0, 0),
-		@Column(name="bio", nullable=false) var bio:String = "",
-		@Column(name="favoriteLang", nullable=false) var favoriteLang: String = "",
-		@Column(name="passWord", nullable=false) var passWord: String,
-		@Column(name="showId", nullable=false) var showId: String,
-		@Column(name="mailAuthorized", nullable=false) var mailAuthorized: Int = 0,
-		@Column(name="mailAddress", nullable=false) var mailAddress: String
+        @Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
+        @Column(name="name", nullable=false) var name: String,
+        @Column(name="birthDay", nullable=false) var birthDay: LocalDateTime? = LocalDateTime.of(1900, 1, 1, 0, 0, 0),
+        @Column(name="bio", nullable=false) var bio:String = "",
+        @Column(name="favoriteLang", nullable=false) var favoriteLang: String = "",
+        @Column(name="passWord", nullable=false) var passWord: String,
+        @Column(name="showId", nullable=false) var showId: String,
+        @Column(name="mailAuthorized", nullable=false) var mailAuthorized: Int = 0,
+        @Column(name="mailAddress", nullable=false) var mailAddress: String
 )
 
 @Entity
 data class Token(
-		@Column(name="token") var token: String,
-		@Column(name="userId") var userId:UUID,
-		@Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null
+        @Column(name="token") var token: String,
+        @Column(name="userId") var userId:UUID,
+        @Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null
 )
 
 @Entity
 data class DevelopExp(
-		@Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
-		@Column(name="userId") var userId: UUID,
-		@Column(name="caption") var caption: String
+        @Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
+        @Column(name="userId") var userId: UUID,
+        @Column(name="caption") var caption: String
 )
 
 @Entity
 data class Follow(
-		@Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
-		@Column(name="userId") var userId: UUID,
-		@Column(name="followingUserId") var followingUserId: UUID
+        @Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
+        @Column(name="userId") var userId: UUID,
+        @Column(name="followingUserId") var followingUserId: UUID
 )
 
 @Entity
 data class MailToken(
-		@Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
-		@Column(name="userId", nullable=false) var userId: UUID,
-		@Column(name="token", nullable=false) var token: String
+        @Id @GeneratedValue(generator="uuid2") @GenericGenerator(name="uuid2", strategy="uuid2") @Column(columnDefinition="BINARY(16)") var id: UUID? = null,
+        @Column(name="userId", nullable=false) var userId: UUID,
+        @Column(name="token", nullable=false) var token: String
 )
 
 data class UserPoster(
-		var name: String,
-		var password: String,
-		var showId: String,
-		@Email var mailAddress: String
+        var name: String,
+        var password: String,
+        var showId: String,
+        @Email var mailAddress: String
 )
 
 data class LoginPoster(
-		var address: String,
-		var password: String
+        var address: String,
+        var password: String
 )
 
 data class AuthPoster(
-		var token: String
+        var token: String
 )
